cancel funtion added, user is now able to cancel edit inputs
The bug responsible was that when edit is pressed the original state
of that users info is stored in a seperate state named original. The filtered
user was aquired through the filter function which returns a shallow copy of
the user in the users state. Later on when the users state was set to a different
state the filtered user the original state changed too because that value was just
pointing to the value in the users state. Fixed with a .find funtion instead of a
filter function.